"Next" & "Check" button fix.
One-line changes were made to the DiagramButtonView and ImageDiagramView files to change the default topic name and add a modifier, on lines 28 and 204 respectively. Additionally, extensive code cleanup was performed in the QuizzView file, and the logic was improved to allow the user to select their preferred option from the Quiz list. The review button is dynamically updated to display the text 'Check' or 'Next'.